refactor: mvc split of color_box.py
