+++ src/tagstudio/qt/mixed/tag_color_label.py
@@ -2,8 +2,6 @@
 # SPDX-License-Identifier: GPL-3.0-only
 
 
-import typing
-
 import structlog
 from PySide6.QtCore import QEvent, Qt, Signal
 from PySide6.QtGui import QAction, QColor, QEnterEvent
@@ -20,10 +18,6 @@ from tagstudio.qt.translations import Translations
 
 logger = structlog.get_logger(__name__)
 
-# Only import for type checking/autocompletion, will not be imported at runtime.
-if typing.TYPE_CHECKING:
-    from tagstudio.core.library.alchemy.library import Library
-
 
 class TagColorLabel(QWidget):
     """A widget for displaying a tag color's name.
@@ -39,11 +33,9 @@ class TagColorLabel(QWidget):
         color: TagColorGroup | None,
         has_edit: bool,
         has_remove: bool,
-        library: "Library | None" = None,
     ) -> None:
         super().__init__()
         self.color = color
-        self.lib: Library | None = library
         self.has_edit = has_edit
         self.has_remove = has_remove
