+++ src/tagstudio/qt/mixed/tag_color_manager.py
@@ -24,8 +24,8 @@ from PySide6.QtWidgets import (
 from tagstudio.core.constants import RESERVED_NAMESPACE_PREFIX
 from tagstudio.core.enums import Theme
 from tagstudio.qt.controllers.field_container_controller import FieldContainer
+from tagstudio.qt.controllers.tag_color_box_controller import TagColorBoxWidget
 from tagstudio.qt.mixed.build_namespace import BuildNamespacePanel
-from tagstudio.qt.mixed.color_box import ColorBoxWidget
 from tagstudio.qt.translations import Translations
 from tagstudio.qt.views.panel_modal import PanelModal
 
@@ -117,7 +117,7 @@ class TagColorManager(QWidget):
             for group, colors in self.driver.lib.tag_color_groups.items():
                 if not group.startswith(RESERVED_NAMESPACE_PREFIX):
                     all_default = False
-                color_box = ColorBoxWidget(group, colors, self.driver.lib)
+                color_box = TagColorBoxWidget(group, colors, self.driver.lib)
                 color_box.updated.connect(
                     lambda: (
                         self.reset(),
